Check function declaration
