+++ src/typechecker/bind.ts
@@ -62,59 +62,6 @@ import { BoundTuple } from "./bound/BoundTuple";
 export const bind = (node: ASTNode, env: TypeEnv, ty?: Type): BoundASTNode => {
   let key, value, synthType;
   switch (node.kind) {
-    case SyntaxNodes.FunctionDeclaration:
-      if (node instanceof FunctionDeclaration) {
-        // gets extended environment from type checker
-        const name = bind(node.name, env, ty!) as BoundIdentifier;
-        const boundParams = node.params.map((p) =>
-          BoundParameter.new(p, p.type ? getType(p.type, env) : Type.any())
-        );
-        const boundBody = bind(
-          node.body,
-          env,
-          (ty! as Type.Function).ret
-        ) as BoundBlock;
-
-        return BoundFunctionDeclaration.new(
-          name,
-          boundParams,
-          boundBody,
-          // ty guaranteed to be a function type passed in from type checker
-          (ty! as Type.Function).ret,
-          node.start,
-          node.end
-        );
-      }
-      // Should never happen
-      throw new Error("WTF, indeed?");
-
-    case SyntaxNodes.Block:
-      if (node instanceof Block) {
-        const exprs = node.expressions;
-        const boundExprs = exprs.map((expr, i, a) => {
-          let type = synth(expr, env);
-
-          if (expr.kind === SyntaxNodes.ReturnStatement || i === a.length - 1) {
-            // ty will be passed in from the type checker
-            if (!isSubtype(type, ty!)) {
-              throw new Error(`Cannot use ${type} as a subtype of ${ty}`);
-            }
-          }
-
-          return bind(expr, env, type);
-        });
-
-        return BoundBlock.new(boundExprs, ty!, node.start, node.end);
-      }
-      throw new Error("WTAF");
-
-    case SyntaxNodes.ReturnStatement:
-      if (!ty) {
-        ty = synth((node as ReturnStatement).expression, env);
-      }
-
-      const boundExpr = bind((node as ReturnStatement).expression, env, ty);
-      return BoundReturnStatement.new(boundExpr, ty!, node.start, node.end);
     default:
       throw new Error(`Cannot bind node of kind ${node.kind}`);
